User + Admin

+++ proiect_daw/Entities/Budget.cs
@@ -8,7 +8,7 @@ namespace proiect_daw.Entities
     public class Budget
     {
         public int Id { get; set; }
-        public float Money { get; set; }
+        public int Money { get; set; }
         public int MovieId { get; set; }
         public Movie Movie { get; set; }
     }
